Sample generic Prepare forms.
The commits provide for sample settings that match the variable names
in AutoDeploy and NSX playbook operations. The intent is to soon move all
variables in playbooks to 'default' variables files and utilize the UI
for all settings other than the defaults.

Note that by default (unless changed in the future, or in Ansible playbook
variables) the answer file containing the Ansible variable file compatible
settings are in /var/lib/supervio/answerfile.yml and that should be included
in future playbooks regarding SuperVIO operations.

Change-Id: Ie1377a83bb75e02fabcf1d3cd5473426a4b5cd76